Fixed issue with crash when running test executable.
Test executable assumes that first environment argument is always taken by the executable name. This means that in instances where the test_main function in rustc/libtest is triggered by other processes where the exectuable name doesn't appear as an environment variable and the environment variables list is empty a panic! is triggered as the env::args() are accessed from slice 1 onwards.

Solution is naturally to add the executable name as the first environment argument!