Now grab functions to cover from files.
As I delve deeper into the dwarf format, I've managed to implement code to grab the functions in the compilation unit. Currently they're just printed out but the next stage is to generate a list of addresses for PTRACE and also start to find the total number of lines which must be exercised by a test.