ASLR would have been enabled before preventing tracing.
Added personality.rs, hopefully it will be implemented in the nix crate in the imminent future (not doing a pull-request as I believe rusty_trap is likely working on one).

ASLR is now disabled before running a test, and tracing still doesn't work but it should be moving be closer.