Lines are now successfully instrumented.
Found that as well as ASLR, Program-Independent Code (PIC) adds an offset to the location the program is placed. This offset I believe is wrote to /proc/ in some location, but I've found it easier to disable it for running coverage.

I've also enabled the flag for rustc to link-dead-code. It could be assumed that if a breakpoint cannot be placed it represents dead code and won't be included in coverage results. However, I'd prefer to for now ensure that every line of code is traceable and remove the false-positives from ptrace.

This also means that breakpoints are now hit, so the next stage is to continue on from the instrumentation point and log the coverage results.

Additionally, updated cargo package from 0.17.0 to 0.18.0