Started implementing code to step through.
Work is done on trying to step through the code. It's close to working but there are unexpected stops which cause it to halt. Need to move ptrace stuff out to some sort of unit to make it easier to use. And potentially break out some of the other larger functions as well.