Fixed breakpoint and altered test launch.
Fixed breakpoint to place the interrupts in the correct place. This was an issue where peeks are byte aligned by instruction addresses aren't necessarily aligned.

Launch tests using a single thread. From running strace on some test executables when there are multiple threads allowed for tests the chance of a panic in a traced process is significantly increased. This should therefore improve stability.