No longer instrument module imports.
Module imports show up in the dwarf debug lines, however they are always ran to import the module and attempting to instrument them was causing intermitent segfaults in the test executable. This seems to be because the address of __test::main was the same as the address as a module import on one of my test projects.

This has added a dependency on the regex crate. For future improvements analysis of code to trace via syntex_syntax crate may be required.