Found ptrace issue.
Previously, there were ptrace issues where the child process was killed. This is because of nuances behind ptrace which can lead to it killing a process.

When ptrace-ing a process the following process should be followed:

* Child signals parent it is stopped
* Parent does pokes and peeks with ptrace, and other operations.
* Parent does a ptrace which continues the child process.
* Parent waits until child signals it is stopped before ptracing again on the child.

Before I was removing the breakpoint, rerunning the instruction without the breakpoint, reenabling the breakpoint and then continuing execution. After I run the instruction without the breakpoint if the instruction to reenable the breakpoint happened before the childs signal state had changed this caused the child to die. This race condition essentially explains how I was getting different results every time I ran the code.

This has facillitated an extra field in the breakpoint called is_running. This shows if the code was running before it hit the breakpoint. If false it means the code wasn't running and instead we're at that instruction pointer as a result of a ptrace single_step.

Refactored the main loop to account for this.

Next step is to identify test entry points and jump us into these, this is because __test::main seems to launch each test on a separate thread even with TEST_THREADS=1. This is likely so it can observe tests for timeouts and kill them if they take too long. It does mess up my ptracing on the functions. So the solution is to jump into each function and run it. Then when leaving the function jump to the next function to avoid a segfault or kill the process.