Now identify LineTypes function and test entry.
Function and test entries are now identified. Test entries are defined as functions within the tests module and functions as all other functions. This assumption won't hold up for users who don't use a tests module despite rust recommendations or users with more complicated tests which require helper functions.

However, for the purposes of preprealpha this is deemed acceptable.