Identifies tests in tests/ folder.
It now identifies all files in the tests/ folder and forces them to a TestEntry as before they were being misidentified as a FunctionEntry (although the path issue I found in objdump wasn't evident :S).

Seems that breakpoints don't work on the functions they call so that's the next step.