Added coveralls_api.
Added coveralls_api crate to the project and stripped out the original attempt at implementation. Changed report to take the config struct instead of a path. This means it will have access to the manifest root and the output directory for any reports.

Also added the coveralls_api options to clap with help text.

Removed collector.rs as it's original intent is being fulfilled by the report crate.