Ptrace updates and improvements.
Have been working on updates to allow tracing across multiple threads and avoid faffing around with registers and deriving correct register start states. This has resulted in a number of ptrace utilities being written to aid in debugging which will eventually be moved to a PR for rust-nix/nix.

Coverage does still not fully work, however it is possible for complete coverage to be reported with some minor fiddles.