Resolves #3
Tests are now traced going from the root of the executable without jumping around. This prevents segfaults when utilising heap allocations in the test or tested functions. This commit adds in changes to find the actual final PC in a test function (as high_pc can be >= the max PC).

Although, it seems only one test in a test binary is ran. This is another issue but this is at least an improvement.