Fixed issues with exiting a test.
Updated breakpoint to take an optional uid (either tid or pid), when the thread caused a SIGTRAP and the breakpoint process used the process id for the test that was what was causing the SEGFAULT issues.

This has now resolved a lot of issues with line coverage and as a result a lot of the errors I've observed in the past have been sorted (at least in all tested instances). As a result I'm pushing a new minor release.