Fixed multiple function in test artefact issue.
When a test function finishes, a ptrace stop event is emitted with the tid. In this instance the underlying pid should be continued as the pid may wish to launch another test function. This is why only the first function was having coverage collected for it. By adding the continue multiple functions can now be covered.

Added a new test project for testing test artefacts with multiple functions in them.

Updated minor version number as this can be considered a patch.