Modified ptrace event loop for more correct handling of tids.
Previous iteration had a continue on a tid in a ptrace event. 90% of the time this is okay as the ptrace event is called when a new thread is created so the thread should be continuable. However, in some instances it seems that the thread may not be in a state where it can respond to a PTRACE_CONT. The more idiomatic (but alas undocumented) approach is to continue from the pid that you're waiting on and then handle the SIGSTOP that's emitted before the tid is ready to respond.

This is now fixed and spurious errors are sorted.