Fixed issue that affected some test restarts.
As the more idiomatic ptrace usage was adopted, the requirement to monitor when the test end was reached was dropped. This has now been removed from the code and has resulted in greater stability.

Additionally, and slightly stranger, if I continue_exec on the exit and get an error I need to handle the next waitpid in order for future test executables to work.