Cleaned up help and verbose logging.
If verbose isn't enabled then the project compile now uses the quiet flag. This can be helpful for projects where the library compile generates a lot of output to stdout.

Need to move the printout of all the lines out of verbose and into one of the reporting methods as a result.