Last bug fix.
Fixed the last bug, this was caused by reading instruction data that contained more than one breakpoint and disabling it by writing the word as it was when the first of the N breakpoints was created. This overwrote the subsequent breakpoints with the original data disabling them.

Solution is to only store the byte overwritten for the interrupt and to do the inverse shift when disabling the breakpoint.

Also, removed debug functionality for reading registers and incremented version to 0.3.0 as this closes the last of the observed issues and the only compiler warnings.