Fixes #6
Added linker argument to disable PIE when linking. In some versions of rustc and some distros with a patched GCC (Ubuntu versions >15.04) PIE can be enabled when PIC is disabled. This renders tarpaulin unable to instrument code and collect coverage.

This commit fixes that issue and pushes the version number to 0.3.6