Fixes #10
DWARF tables changed with rustc 1.19 and source files in the project were referred to with relative paths instead of absolute, this caused the logic to determine which files should be covered to fail.

Tarpaulin now looks for whether the path is relative or absolute. If relative it prefixes the path with the current working directory.