Removed linker flag.
Not every linux system is guaranteed to have the same linker flags available. And not every user is guaranteed to be using the gcc linker, some may choose to use alternative linkers. To avoid flags which may render tarpaulin unusable for a subset of users no linker flags are passed and instead a warning is printed if tarpaulin is unable to instrument the code.