Check the linker used by rustc to determine if "-no-pie" should be passed.
This implementation gets the configured linker for the target if one has been
set using the same meathod as Cargo. That can be found in `scrape_build_config`
and `scrape_target_config` in "src/cargo/ops/cargo_compile.rs" in Cargo.

Sadly we cannot query `rustc` for its default linker for a target without
nightly-only shenanigans so we have to make some assuptions. For Linux (and most
everything that isn't Windows) it is fair to assume the default linker is `cc`
and that `cc` is GCC based. Even if it is not, if the call to the linker to get
its version information fails to does not contain "--enable-default-pie" no
flags are set.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>