Fixes #12
Added the forward option to fix issue #12 involving tarpaulin stealing signals.

The forward option means that in the event of tarpaulin receiving an unexpected signal it will forward it onto the test. This is for tests which test functionality involving unix signals. This still won't allow test which make use of SIGTRAP or SIGSTOP from functioning with tarpaulin as it is not possible to determine which signals are for tarpaulin and which are for testing.