Continued work on #21
Multithreaded breakpoint now restores the instruction that was there when the
breakpoint was previously hit. This prevents the instruction being overwritten
with a breakpoint and the program getting stuck.

Unfortunately it still seems to get stuck on a breakpoint on occasion, so to
combat this a breakpoint timeout was implemented so if it is stuck on the same
breakpoint for more than 100 SIGTRAPS it will disable the breakpoint for the rest
of the coverage.

An occasional segfault was detected in the test so detection and killing of it has been implemented.