Unused generic code is now ignored.
After adding all the addresses and lines from the DWARF tables the cover hashset is now iterated to add any lines that should be covered that may not be in the binary to the tracerdata. These are given the line type of UnusedGeneric as currently it only affects generic code which never gets a concrete type put in.