Fixed issues with deduplicating generics.
Following generic function deduplication, one concrete implementation would be instrumented and the others wouldn't. This meant erroneous coverage could be reported in some instances. Deduplication now happens for every line program sequence. This is because specialisations are given a separate lineprogram sequence.

This results in TracerData with duplicate path:line combos so an additional deduplication step was added before report generation to merge the specialisation results. This stage also added sorting of the elements so now line numbers are always in order which is a nice enhancement!