Now ignore comments in unused functions/methods
Previous if a function or method was unused it's entire span was added to the "coverable" map and ignorable line rules trimmed out the erroneous lines. Except for comments which don't show up in the AST. Added regex to parse for single or multiline comments and added rules to ignore them.