Deleted a bunch from test_loader
Earlier tarpaulin worked on the assumption of jumping between different tests to avoid segfault issues. As these issues were caused by not handling ptrace events and threading and are no longer a thing a lot of this code became unnecessary. This is general housekeeping removing said code as I find it.