Implemented fix for missing lines
Fix for missing lines is that traces may have to be added for each compilation unit to handle different addresses. Solved this by in the test loader moving the results from each CU into a TraceMap and merging it with a master TraceMap which is then returned. Also, adding a call to dedup before reporting results then collapses these extra traces into a single trace.