Improved tracing of more complex lines.
If a line appears in multiple line number programs, the lowest address for that line is traced for every program it appears in. Whereas before, the lowest address for a line in the compilation unit was traced. This adds extra redundancy and improves tracing of more complicated features such as `?` etc.