Improved coverage of chained calls
Improved coverage of chained calls potentially solving the pesky `?` issue. Now non-sequential instances of a location in a line number program are considered as trace points. This will have an impact on memory usage because there's more trace points so consider in future looking for whether an instruction is worth tracing. There must be other indications than DWARF.