actions: Don't fail-fast in the matrix build
By default, GitHub Actions "fails fast;" if one of the builds fails, the
rest are cancelled, presumably to conserve build minutes.

Since both (i) GitHub Actions is free and unlimited for OSS projects,
and (ii) nightly builds are not necessarily a great indication of
whether or not a given changeset is actually broken since nightly does
break from time to time, it seems wise to prevent failures in e.g. the
nightly build from stopping the entire workflow.

Essentially, this commit allows stable/beta builds to continue if nightly
fails, but the same does also apply in the other direction.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>