Bump required clap version to 2.33.0 in Cargo.toml
Cargo.toml currently specifies a minimum clap version of 2.31.2. However, this is an ol
version from March 2018 that cargo-tarpaulin 0.10.0 does not actually compile against.

This is because of a macro issue for Rust 2018 in Clap 2.31.2, where the arg_enum macro
emits a call to the _clap_count_args macro expecting it to be in scope. Since tarpaulin
does not `use clap::macros::_clap_count_args` this is not true, and compilation fails.

This was fixed in this Dec 2018 commit to `clap`: https://github.com/clap-rs/clap/commit/cfbcb93f94fe8402054945e946bd51ed10b57326#diff-e1576943a5907b15388ab726037ebdb3 which is
included in 2.33.0

Having the old version in Cargo.toml  breaks using cargo-tarpaulin as a library, as in the
`runtime_macros` crate, which I was attempting to upgrade to the newest version of
tarpaulin, as in this case the Cargo.toml version is consulted. It has not caused any
issues for using it as a binary as in this case Cargo.lock is consulted which already
lists 2.33.0.

This commit updates the version in Cargo.toml to match the version that tarpaulin
is actually built against and compatible with.