Add address to SourceLocation struct
Becasue address is used to determine if a line has been covered, this
commit's changes add address to the key used for maps of TracerData.

The gimli object used to describe line and path can be associated with multiple addresses
which resulted in scenarios where lines that _were_ covered were missed
in the reports (because a TracerData associated with line and path, but
not the expected address, got to the map first).

In this commit, a test was added to the `tests/data/matches` crate to illustrate a
scenario that shows a line that was previously uncovered in reporting,
is now covered.