Add match expr test and run in test suite
This commit adds an extra scenario to the suite of match expression
tests. It was added because the changes in commit b8b427cdde6f996df20d4dfbf95a4b19d24a76da
add support for it.

This commit also adds the matches tests to those run from
`tests/mod.rs`. I'm not actually sure why it was omitted from that file
though so maybe that's the wrong thing to do :grimacing: