Integrate should_panic support and clean doctest folder
should_panic status of a test is now propagated up and taken into account when collecting the return code. Also during testing it was found that target/doctest wasn't cleaned and old doctests would be ran if a name clash hadn't replaced the file. Doctests are now cleaned between runs