Fix doctests for nightly 1.50
So 1.50 changed the doctest naming scheme to add the package name into
it. This seems like an alright solution? It might break for workspaces
but some of my assumptions broke anyway and there were possible
conflicts... If this causes the current tests to pass it should still
work for 1.49 so  I'm kinda okay with that