Fixes #846
When an outer attribute is used on a file i.e.
`#![cfg(not(tarpaulin_include))] to exclude it, if this file is a
lib.rs, mod.rs or <module_name>.rs all other files in that module should
be excluded. Now when one of these attributes is picked up tarpaulin
will check if any files covered by it have been processed and if so wipe
all the analysis and set them to ignore all, then for future files that
match the module path it will ignore them as well.