Exclude files already when walking source tree
It is known that `tarpaulin` will look for coverage data in vendored
crates if `cargo-vendor` is used (#210). It is reasonably simple to
exclude the vendored crates though the `--exclude-files` option.
However, all of the sources is still collected even if they are not
treated as covered. This is quite obvious when using the `Html` report
format for example which becomes huge because it contains the crate
sources as well as all the vendored crates' source files.

To avoid this, apply the excluded files filter at the earliest point,
i.e., when walking the crate source tree and finding all the rust source
files.