Remove sched_setaffinity call
So rustc test runner now uses the affinity to set default number of
threads, and tarpaulin's previous method to force it to one core caused
that to make all tests run single threaded. That was introduced to try
and solve a ptrace issue but I feel _strongly_ that it isn't actually
necessary and just minimised the chance of hitting another issue that is
since fixed. Given gdb doesn't use this sys call I feel fairly confident
in removing it.