Revamp --follow-exec
This does a few changes to --follow-exec, to stop it stalling the test processes and capture more exec'ed processes. This works in two parts:

 posix_spawn to start an executable works by doing a vfork and then exec in the child, this revamps the vfork handling to catch exec calls.

When a fork is traced in ptrace, the parent is halted until the child exits or returns. When we do `Command::new().spawn()` and then want to issue a signal to the executable under test this would result in a deadlock when using --follow-exec. This is because the parent needs to be explicitly continued. However, it cannot be continued until after the child exec completes or this causes _issues_. This PR adds the parent continue to a pending actions queue and then in the next iteration of the state machine it continues the parent so it can run independently to the child and issue signals/respond to it so a wider range of integration tests work.

There was one spurious segfault (~4% of rusn on my machine, >90% on CI) in nightly, which seems to be the intermingling of one process exiting and another continuining in the same update. To solve this --test-threads=1 is the default for --follow-exec runs now.

This also refines a lot of the event log stuff adding it better fork support and markers to show where update iterations happen. This has proven very useful for debugging.