Fix a bunch of cargo config stuff...
So the previous way cargo-tarpaulin handled the config files was a bit
lazy. It only looked in the project directory and the home directory. Now I
follow the actual compiler behaviour and go from project to root
looking for config files then the one in the home directory merging
things as we go...

This also grabs the env vars from the config files and applies them when
the metadata is gathered, project built and project ran!